+++ easyarchive-starter/src/main/java/com/openquartz/easyarchive/starter/model/vo/ArchiveGroupItemByIdVO.java
@@ -34,7 +34,7 @@ public class ArchiveGroupItemByIdVO {
 
     private Integer stepCount;
 
-    private Integer stepRounds;
+    private Long stepRounds;
 
     private Integer pauseMs;
